+++ cardinal-components-world/src/main/java/org/ladysnake/cca/mixin/world/common/DataFixTypesMixin.java
@@ -25,7 +25,7 @@ package org.ladysnake.cca.mixin.world.common;
 import com.mojang.datafixers.DataFixer;
 import com.mojang.serialization.Dynamic;
 import net.minecraft.util.datafix.DataFixTypes;
-import org.ladysnake.cca.internal.world.ComponentPersistentState;
+import org.ladysnake.cca.internal.world.CcaSaveData;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -39,7 +39,7 @@ public class DataFixTypesMixin {
         cancellable = true
     )
     private <T> void cancelIfCca(DataFixer dataFixer, Dynamic<T> dynamic, int oldVersion, int newVersion, CallbackInfoReturnable<Dynamic<T>> cir) {
-        if (ComponentPersistentState.LOADING.get()) {
+        if (CcaSaveData.LOADING.get()) {
             cir.setReturnValue(dynamic);
         }
     }